changed abv columns to database type double to maintain precision

+++ src/main/resources/db.sql
@@ -5,8 +5,8 @@ create table beer_style (
   description varchar(4000),
   ibu_min integer,
   ibu_max integer,
-  abv_min numeric,
-  abv_max numeric,
+  abv_min double,
+  abv_max double,
   brewerydb_create_date timestamp,
   unique (brewerydb_id)
 );
@@ -17,7 +17,7 @@ create table beer (
   brewerydb_id varchar(8),
   brewerydb_brewery_id varchar(8),
   description varchar(4000),
-  abv numeric,
+  abv double,
   ibu integer,
   brewerydb_create_date timestamp,
   style_fk integer,
